project/requirements.txt
Django>=3.0,<4.0
django-environ
djangorestframework
django-cors-headers
mysqlclient
